Work around issues with bouncing zoom transitions
When using the zoom buttons, a transitions is used to go from one zoom
level to the other smoothly. Internally in d3, the zoom animation isn't
using a linear interpolation, but uses the d3.interpolateZoom function.
This is a non-linear function that will sometimes overshoot the initial
or target zoom level by a large marging, leading to bouncing and very
unintuitive behavior, e.g. because the animation will first zoom out on
a zoom-in event.

This patch works around the issue by globally replacing
d3.interpolateZoom with the linearly interpolating d3.interpolateArray.